Make error messages better if getTBAQuery fails
Prevent ReferenceError if query fails and cached data is not available
for getTBAQuery (before it would give a reference error trying to get
cached.data even if cached was undefined, so instead of getting a
'Error: [status code] received from TBA attempting to process query' it
would result in a reference error.

Also create an alert box with the error if no event information could be
loaded at all so what is wrong is more obvious if it ever happens.