Move openURLKey to after event markers are added
getTBAQuery is asynchronous since it accesses TBA, so we have to
open the url requested key (?key=eventCode) in the callback after all
the events are added, instead of after the request function is called.

If this is not done, the event markers probably won't exist when
openURLKey is called, so an event marker specified in the url will
not be opened, it will just be ignored.