Make links open in a new tab
Make all links open in a new tab by default. Before this, team
websites and TBA pages opened in the same tab as the
map, so in order to go back to the map the user had to hit the
back button, and the map had to reload.

I feel like it makes more sense for team websites and TBA pages to open in a new tab by default, so the user can go back to the map more easily and without it reloading.

This also affects links on the About menu, which I think is
fine. If any other links are added that should open in the
same tab, it can be manually specified using the <a> tag's
`target` attribute.